requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
EasyProcess==1.1
entrypoint2==1.1
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.7
pywhatkit==5.4
redmail==0.6.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
Werkzeug==3.0.0
wikipedia==1.4.0
